feat: add validation for users

+++ src/users/users.controller.ts
@@ -6,10 +6,18 @@ import {
   Put,
   Param,
   Delete,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { IsUUID } from 'class-validator';
+
+export class FindOneParams {
+  @IsUUID(4)
+  id: string;
+}
 
 @Controller('users')
 export class UsersController {
@@ -28,17 +36,30 @@ export class UsersController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param() { id }: FindOneParams) {
+    if (!this.usersService.getById(id))
+      throw new HttpException("User don't found", HttpStatus.NOT_FOUND);
     return this.usersService.getById(id);
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  update(@Param() { id }: FindOneParams, @Body() updateUserDto: UpdateUserDto) {
+    const user = this.usersService.getById(id);
+    if (!user)
+      throw new HttpException("User don't found", HttpStatus.NOT_FOUND);
+    if (user.password !== updateUserDto.oldPassword)
+      throw new HttpException(
+        'Old password is incorrect',
+        HttpStatus.FORBIDDEN,
+      );
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param() { id }: FindOneParams) {
+    const user = this.usersService.getById(id);
+    if (!user)
+      throw new HttpException("User don't found", HttpStatus.NOT_FOUND);
     return this.usersService.remove(id);
   }
 }
